feature: converted the Alexa SDK typing to a fully ambient module
Conversion to fully ambient means you don't have to do any includes to use it.
Also renamed the module from AlexaSDK to Alexa, leaning on TypeScript merging to share the symbol with the global Alexa instance.
Fixed declaration of the second argument to sendMessage to be a callback instead of just the callback argument.